# Conflicts:
#	gui/src/app/components/map-container/map-container.component.ts
#	gui/src/app/pages/main-page/main-page.component.html
#	gui/src/app/pages/main-page/main-page.component.ts

+++ backend/src/data/mod.rs
@@ -21,7 +21,7 @@ pub struct SightsConfig {
 pub struct CategoryTagMap {
     category: String,
     opening_hours: String,
-    duration_of_stay_minutes : usize,
+    duration_of_stay_minutes : i64,
     tags: Vec<Tag>
 }
